Refactor: AdminChallengeOpen 템플릿 분리
