+++ src/pages/AdminChallengeOpen/template/style.ts
